# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Card API
  description: API for managing bank cards
  contact:
    name: Junior Dev Team
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/users/{id}:
    get:
      tags:
        - User API
      summary: Получить пользователя по ID
      description: Возвращает информацию о пользователе по его ID. Доступно только администратору.
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о пользователе успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
      security:
        - bearerAuth: []
  /api/users/self:
    get:
      tags:
        - User API
      summary: Получить текущего пользователя
      description: Возвращает информацию о текущем авторизованном пользователе.
      operationId: getCurrentUser
      responses:
        '200':
          description: Информация о пользователе успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
      security:
        - bearerAuth: []
  /api/users/all:
    get:
      tags:
        - User API
      summary: Получить всех пользователей
      description: Возвращает список всех пользователей с пагинацией. Доступно только администратору.
      operationId: getAllUsers
      parameters:
        - name: pageable
          in: query
          description: Параметры пагинации (page, size, sort)
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список пользователей успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserDTO'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserDTO'
        '403':
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserDTO'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        email:
          type: string
          description: Email пользователя
          example: testuser@example.com
        fullName:
          type: string
          description: Полное имя пользователя
          example: John Doe
        role:
          type: string
          description: Роль пользователя
          example: USER
          enum:
            - ROLE_USER
            - ROLE_ADMIN
      description: DTO для представления информации о пользователе
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageUserDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDTO'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT